requirements.txt
fpdf==1.7.2
pandas==1.3.3
google-generativeai
python-dotenv
PyPDF2
docx
pinecone-client
llama-index